Switch using to listen/resume from wait/give.
