+++ api/package.json
@@ -39,7 +39,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.16.3",
-    "@findy-network/findy-common-ts": "^0.1.2",
+    "@findy-network/findy-common-ts": "^0.1.3",
     "aws-sdk": "^2.1025.0",
     "axios": "^0.24.0",
     "config": "^3.3.6",
